Fix build issue

+++ android-exoplayer/src/main/java/com/brentvatne/exoplayer/ReactExoplayerView.java
@@ -451,7 +451,7 @@ class ReactExoplayerView extends FrameLayout implements
             }
             long usedMemory = runtime.totalMemory() - runtime.freeMemory();
             long freeMemory = runtime.maxMemory() - usedMemory;
-            long RESERVE_MEMORY = 0.2 * runtime.maxMemory();
+            long RESERVE_MEMORY = (long)0.2 * runtime.maxMemory();
             if (RESERVE_MEMORY < freeMemory) {
                 // We have less than 20% of app memory available to use, we want to keep that in reserve
                 Log.w("ExoPlayer Warning", "Free memory reached <20%, forcing garbage collection");
